Save media store differently than data store
The media store is a large blob of JSON that maps media links to avoid
having to reupload avatars constantly. This JSON blob can quickly grow
to rather obscene sizes, so it's now stored as zipped and base64 encoded
data, to avoid limits on account_data sizes.

This is just a temporary fix until a better design is inished